+++ core/src/main/java/cucumber/runtime/converters/ConverterWithEnumFormat.java
@@ -1,6 +1,6 @@
 package cucumber.runtime.converters;
 
-import com.thoughtworks.xstream.converters.ConversionException;
+import cucumber.runtime.xstream.converters.ConversionException;
 
 import java.text.FieldPosition;
 import java.text.Format;