+++ core/src/main/java/cucumber/table/xstream/ComplexTypeWriter.java
@@ -1,7 +1,5 @@
 package cucumber.table.xstream;
 
-import cucumber.runtime.CucumberException;
-
 import java.util.ArrayList;
 import java.util.List;